requirements.txt
cairocffi==0.9.0
CairoSVG==2.2.1
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
colorama==0.4.1
cssselect2==0.2.1
defusedxml==0.5.0
Django==2.1.3
django-ckeditor==5.6.1
django-js-asset==1.1.0
djangorestframework==3.9.0
html5lib==1.0.1
httpie==1.0.2
idna==2.8
oauthlib==2.1.0
Pillow==5.3.0
psycopg2==2.7.6.1
pycparser==2.19
Pygments==2.3.1
PyJWT==1.7.1
Pyphen==0.9.5
python3-openid==3.1.0
pytz==2018.7
requests==2.21.0
requests-oauthlib==1.0.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==2.0.0
tinycss2==0.6.1
urllib3==1.24.1
WeasyPrint==43
webencodings==0.5.1
XlsxWriter==1.1.2